♻️  Refact: 중복 코드 및 위치 모델 폴더링
